requirements.txt
aiohttp>=3.6.2
aiohttp-sse-client>=0.1.7
beautifulsoup4>=4.9.1
Brotli>=1.0.9
certifi>=2020.6.20
chardet>=3.0.4
click>=7.1.2
dash>=1.16.2
dash-bootstrap-components>=0.10.6
dash-core-components>=1.12.1
dash-html-components>=1.1.1
dash-renderer>=1.8.2
dash-table>=4.10.1
dash_extensions>=0.0.31
Flask>=1.1.2
Flask-Caching>=1.9.0
Flask-Compress>=1.5.0
future>=0.18.2
gunicorn>=20.0.4
idna>=2.10
git+https://github.com/addisonlynch/iexfinance.git#egg=iexfinance
itsdangerous>=1.1.0
Jinja2>=2.11.2
MarkupSafe>=1.1.1
numpy>=1.19.2
pandas>=1.1.2
plotly>=4.10.0
python-dotenv>=0.14.0
redis>=3.5.3
requests>=2.24.0
retrying>=1.3.3
six>=1.15.0
sseclient>=0.0.27
urllib3>=1.25.10
Werkzeug>=1.0.1
